+++ src/interface/session.ts
@@ -4,12 +4,35 @@ import { IncomingHttpHeaders, OutgoingHttpHeaders } from "http";
  * Represents various settings for HTTP/2 (h2) protocol.
  */
 export interface H2Settings {
-  HEADER_TABLE_SIZE?: number; // The size of the header compression table used in HTTP/2.
-  ENABLE_PUSH?: boolean; // Indicates whether server push is enabled in HTTP/2.
-  MAX_CONCURRENT_STREAMS?: number; // The maximum number of concurrent streams allowed in HTTP/2.
-  INITIAL_WINDOW_SIZE?: number; // The initial window size for HTTP/2 data frames.
-  MAX_FRAME_SIZE?: number; // The maximum size of HTTP/2 frames.
-  MAX_HEADER_LIST_SIZE?: number; // The maximum size of header list in HTTP/2.
+  /**
+   * The size of the header compression table used in HTTP/2.
+   */
+  HEADER_TABLE_SIZE?: number;
+
+  /**
+   * Indicates whether server push is enabled in HTTP/2.
+   */
+  ENABLE_PUSH?: boolean;
+
+  /**
+   * The maximum number of concurrent streams allowed in HTTP/2.
+   */
+  MAX_CONCURRENT_STREAMS?: number;
+
+  /**
+   * The initial window size for HTTP/2 data frames.
+   */
+  INITIAL_WINDOW_SIZE?: number;
+
+  /**
+   * The maximum size of HTTP/2 frames.
+   */
+  MAX_FRAME_SIZE?: number;
+
+  /**
+   * The maximum size of header list in HTTP/2.
+   */
+  MAX_HEADER_LIST_SIZE?: number;
 }
 
 /**
@@ -56,17 +79,32 @@ export type CertCompressionAlgo = "zlib" | "brotli" | "zstd";
  * Represents parameters for setting priority of HTTP/2 streams.
  */
 export interface PriorityParam {
-  streamDep: number; // The stream dependency for priority.
-  exclusive: boolean; // Indicates whether the priority is exclusive.
-  weight: number; // The weight of the priority.
+  /**
+   * The stream dependency for priority.
+   */
+  streamDep: number;
+  /**
+   * Indicates whether the priority is exclusive.
+   */
+  exclusive: boolean;
+  /**
+   * The weight of the priority.
+   */
+  weight: number;
 }
 
 /**
  * Represents priority frames for HTTP/2 streams.
  */
 export interface PriorityFrames {
-  streamID: number; // The stream ID for which priority is set.
-  priorityParam: PriorityParam; // The parameters for priority settings.
+  /**
+   * The stream ID for which priority is set.
+   */
+  streamID: number;
+  /**
+   * The parameters for priority settings.
+   */
+  priorityParam: PriorityParam;
 }
 
 /**
@@ -177,47 +215,180 @@ export enum ClientIdentifier {
  * Represents options for creating a session.
  */
 export interface SessionOptions {
-  sessionId?: string; // The unique identifier for the session.
-  headers?: OutgoingHttpHeaders; // The headers for the session.
-  proxy?: string; // The proxy server to use for the session.
-  clientIdentifier?: ClientIdentifier; // The client identifier for the session.
-  ja3string?: string; // The JA3 string for the session.
-  h2Settings?: H2Settings; // The settings for HTTP/2.
-  h2SettingsOrder?: (keyof H2Settings)[]; // The order of HTTP/2 settings.
-  supportedSignatureAlgorithms?: SupportedSignatureAlgorithms[]; // The supported signature algorithms.
-  supportedVersions?: SupportedVersions[]; // The supported versions of TLS.
-  keyShareCurves?: KeyShareCurves[]; // The supported elliptic curves for key exchange.
-  certCompressionAlgo?: CertCompressionAlgo; // The certificate compression algorithm.
-  pseudoHeaderOrder?: PseudoHeaderOrder[]; // The order of pseudo headers.
-  connectionFlow?: number; // The connection flow for the session.
-  priorityFrames?: PriorityFrames[]; // The priority frames for HTTP/2 streams.
-  headerOrder?: string[]; // The order of headers.
-  alpnProtocols?: string[]; // The ALPN protocols.
-  alpsProtocols?: string[]; // The ALPS protocols.
-  headerPriority?: PriorityParam; // The priority of headers.
-  randomTlsExtensionOrder?: boolean; // Whether to randomize TLS extension order.
-  forceHttp1?: boolean; // Whether to force HTTP/1 protocol.
-  debug?: boolean; // Whether to enable debugging.
-  insecureSkipVerify?: boolean; // Whether to skip SSL certificate verification.
+  /**
+   * The unique identifier for the session.
+   */
+  sessionId?: string;
+
+  /**
+   * The headers for the session.
+   */
+  headers?: OutgoingHttpHeaders;
+
+  /**
+   * The proxy server to use for the session.
+   */
+  proxy?: string;
+
+  /**
+   * The client identifier for the session.
+   */
+  clientIdentifier?: ClientIdentifier;
+
+  /**
+   * The JA3 string for the session.
+   */
+  ja3string?: string;
+
+  /**
+   * The settings for HTTP/2.
+   */
+  h2Settings?: H2Settings;
+
+  /**
+   * The order of HTTP/2 settings.
+   */
+  h2SettingsOrder?: (keyof H2Settings)[];
+
+  /**
+   * The supported signature algorithms.
+   */
+  supportedSignatureAlgorithms?: SupportedSignatureAlgorithms[];
+
+  /**
+   * The supported versions of TLS.
+   */
+  supportedVersions?: SupportedVersions[];
+
+  /**
+   * The supported elliptic curves for key exchange.
+   */
+  keyShareCurves?: KeyShareCurves[];
+
+  /**
+   * The certificate compression algorithm.
+   */
+  certCompressionAlgo?: CertCompressionAlgo;
+
+  /**
+   * The order of pseudo headers.
+   */
+  pseudoHeaderOrder?: PseudoHeaderOrder[];
+
+  /**
+   * The connection flow for the session.
+   */
+  connectionFlow?: number;
+
+  /**
+   * The priority frames for HTTP/2 streams.
+   */
+  priorityFrames?: PriorityFrames[];
+
+  /**
+   * The order of headers.
+   */
+  headerOrder?: string[];
+
+  /**
+   * The ALPN protocols.
+   */
+  alpnProtocols?: string[];
+
+  /**
+   * The ALPS protocols.
+   */
+  alpsProtocols?: string[];
+
+  /**
+   * The priority of headers.
+   */
+  headerPriority?: PriorityParam;
+
+  /**
+   * Whether to randomize TLS extension order.
+   */
+  randomTlsExtensionOrder?: boolean;
+
+  /**
+   * Whether to force HTTP/1 protocol.
+   */
+  forceHttp1?: boolean;
+
+  /**
+   * Whether to enable debugging.
+   */
+  debug?: boolean;
+
+  /**
+   * Whether to skip SSL certificate verification.
+   */
+  insecureSkipVerify?: boolean;
+
+  /**
+   * The timeout duration for each request, in milliseconds.
+   */
   timeout?: number;
+
+  /**
+   * If true, IPv4 is disabled for the session requests.
+   */
+  disableIPV4?: boolean;
+
+  /**
+   * If true, IPv6 is disabled for the session requests.
+   */
+  disableIPV6?: boolean;
 }
 
 /**
  * Represents base options for making HTTP requests, excluding the body.
  */
 export interface BaseRequestOptions {
-  headers?: OutgoingHttpHeaders; // The headers for the request.
-  redirect?: boolean; // Whether to follow redirects.
-  additionalDecode?: boolean; // Whether to perform additional decoding.
-  proxy?: string; // The proxy server to use for the request.
-  cookies?: Record<string, any>; // Cookies for the request.
+  /**
+   * The headers for the request.
+   */
+  headers?: OutgoingHttpHeaders;
+
+  /**
+   * Whether to follow redirects.
+   */
+  redirect?: boolean;
+
+  /**
+   * Whether to perform additional decoding.
+   */
+  additionalDecode?: boolean;
+
+  /**
+   * The proxy server to use for the request.
+   */
+  proxy?: string;
+
+  /**
+   * Cookies for the request.
+   */
+  cookies?: Record<string, any>;
+
+  /**
+   * If true, indicates the response is in binary format, like base64-encoded images.
+   */
+  byteResponse?: boolean;
+
+  /**
+   * Used to override the Host header, typically needed when making requests directly to an IP address.
+   */
+  hostOverride?: string | null;
 }
 
 /**
  * Represents options for making HTTP requests that may include a body.
  */
 export interface RequestOptions extends BaseRequestOptions {
-  body?: any; // The request body.
+  /**
+   * Request body.
+   */
+  body?: any;
 }
 
 /**
@@ -259,26 +430,88 @@ export interface HeadRequestOptions extends BaseRequestOptions {}
  *  Represents options for making HTTP requests.
  */
 export interface fetchOptions {
-  method?: Methods; // Request method.
-  headers?: OutgoingHttpHeaders; // The headers for the request.
+  /**
+   * Request method.
+   */
+  method?: Methods;
+
+  /**
+   * The headers for the request.
+   */
+  headers?: OutgoingHttpHeaders;
+
+  /**
+   * The body of the request.
+   */
   body?: any;
-  redirect?: boolean; // Whether to follow redirects.
-  additionalDecode?: boolean; // Whether to perform additional decoding.
-  proxy?: string; // The proxy server to use for the request.
-  cookies?: Record<string, any>; // Cookies for the request.
+
+  /**
+   * Whether to follow redirects.
+   */
+  redirect?: boolean;
+
+  /**
+   * Whether to perform additional decoding.
+   */
+  additionalDecode?: boolean;
+
+  /**
+   * The proxy server to use for the request.
+   */
+  proxy?: string;
+
+  /**
+   * Cookies for the request.
+   */
+  cookies?: Record<string, any>;
+
+  /**
+   * Additional session options.
+   */
   options?: SessionOptions;
 }
 
 /**
- * Represents a response from shared library.
+ * Represents a response from the shared library.
  */
 export interface TlsResponse {
-  id: string; // The unique identifier for the response.
-  body: string; // The body of the response.
-  cookies: Object; // Cookies set in the response.
-  headers: IncomingHttpHeaders; // Headers included in the response.
-  sessionId: string; // The session ID associated with the response.
-  status: number; // The HTTP status code of the response.
-  target: string; // The target URL of the response.
-  usedProtocol: string; // The protocol used for the response.
+  /**
+   * The unique identifier for the response.
+   */
+  id: string;
+
+  /**
+   * The body of the response.
+   */
+  body: string;
+
+  /**
+   * Cookies set in the response.
+   */
+  cookies: Record<string, string>;
+
+  /**
+   * Headers included in the response.
+   */
+  headers: IncomingHttpHeaders;
+
+  /**
+   * The session ID associated with the response.
+   */
+  sessionId: string;
+
+  /**
+   * The HTTP status code of the response.
+   */
+  status: number;
+
+  /**
+   * The target URL of the response.
+   */
+  target: string;
+
+  /**
+   * The protocol used for the response.
+   */
+  usedProtocol: string;
 }