+++ src/lib/Cookie.ts
@@ -14,39 +14,36 @@ export class Cookies extends CookieJar {
   }
 
   /**
-   * Checks and sets cookies for a given URL.
-   * If no cookies are provided, it fetches the cookies from the URL.
+   * Fetches all cookies and organizes them by URL.
    *
-   * @param cookies - An object containing cookies as key-value pairs.
-   * @param url - The URL for which cookies are to be set.
-   * @returns An object containing cookies as key-value pairs.
+   * This method serializes cookies and groups them by their domain and path,
+   * constructing a URL as the key and an object of cookies as key-value pairs.
    *
-   * @example
-   * check({ 'cookie1': 'value1', 'cookie2': 'value2' }, 'http://example.com')
-   */
-  public async check(cookies: Record<string, string> | null, url: string) {
-    if (!cookies) {
-      return this.fetchJSON(url);
-    }
-
-    for (const [key, value] of Object.entries(cookies)) {
-      this.setCookieSync(`${key}=${value}`, url);
-    }
-
-    return this.fetchJSON(url);
-  }
-
-  /**
-   * Fetches the cookies for a given URL as a string.
-   *
-   * @param url - The URL from which cookies are to be fetched.
-   * @returns A string containing the cookies.
+   * @returns An object where keys are URLs and values are objects containing cookies as key-value pairs.
    *
    * @example
-   * fetchString('http://example.com')
+   *  {
+   *    "https://example.com/": {
+   *      "cookie1": "value1",
+   *      "cookie2": "value2"
+   *    },
+   *    "https://anotherdomain.com/": {
+   *      "cookieA": "valueA",
+   *      "cookieB": "valueB"
+   *    }
+   *  }
    */
-  public fetchString(url: string) {
-    return this.getCookieStringSync(url);
+  public fetchAllCookies() {
+    const cookies = this.serializeSync().cookies;
+
+    return cookies.reduce((acc, cookie) => {
+      const url = `https://${cookie.domain}${cookie.path}`;
+      if (!acc[url]) {
+        acc[url] = {};
+      }
+      acc[url][cookie.key] = cookie.value;
+      return acc;
+    }, {});
   }
 
   /**
@@ -58,7 +55,7 @@ export class Cookies extends CookieJar {
    * @example
    * fetchJSON('http://example.com')
    */
-  public fetchJSON(url: string): Record<string, string> {
+  public fetchCookiesObject(url: string): Record<string, string> {
     return this.getCookiesSync(url).reduce((acc, cookie) => {
       acc[cookie.key] = cookie.value;
       return acc;
@@ -75,7 +72,7 @@ export class Cookies extends CookieJar {
    * @example
    * fetchSequence('http://example.com')
    */
-  public fetchSequence(url: string) {
+  public fetchCookiesList(url: string) {
     return this.getCookiesSync(url).map((cookie) => ({
       name: cookie.key,
       value: cookie.value,
@@ -83,7 +80,27 @@ export class Cookies extends CookieJar {
   }
 
   /**
-   * Merges the provided cookies with the existing cookies for a given URL.
+   * Checks and sets cookies for a given URL.
+   *
+   * @param cookies - An object containing cookies as key-value pairs.
+   * @param url - The URL for which cookies are to be set.
+   * @returns An object containing cookies as key-value pairs.
+   *
+   * @example
+   * check({ 'cookie1': 'value1', 'cookie2': 'value2' }, 'http://example.com')
+   */
+  public syncCookies(cookies: Record<string, string>, url: string) {
+    if (!cookies) return this.fetchCookiesObject(url);
+
+    for (const [key, value] of Object.entries(cookies)) {
+      this.setCookieSync(`${key}=${value}`, url);
+    }
+
+    return this.fetchCookiesObject(url);
+  }
+
+  /**
+   * Merges the provided cookies with the existing cookies for a given URL according to request payload.
    *
    * @param cookies - An object containing cookies as key-value pairs.
    * @param url - The URL for which cookies are to be set.
@@ -92,10 +109,10 @@ export class Cookies extends CookieJar {
    * @example
    * merge({ 'cookie1': 'value1', 'cookie2': 'value2' }, 'http://example.com')
    */
-  public merge(cookies: Record<string, string>, url: string) {
+  public mergeCookies(cookies: Record<string, string>, url: string) {
     for (const [key, value] of Object.entries(cookies)) {
       this.setCookieSync(`${key}=${value}`, url);
     }
-    return this.fetchSequence(url);
+    return this.fetchCookiesList(url);
   }
 }