+++ typings/src/lib/index.d.ts
@@ -1,0 +1,4 @@
+export * from "./Session";
+export * from "./Response";
+export * from "./Cookie";
+export * from "./requests";