+++ typings/src/utils/koffi.d.ts
@@ -1,0 +1,2 @@
+import { koffiLoad } from "../interface/koffi";
+export declare function load(): Promise<koffiLoad>;