+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# 1.1.4 (July 29, 2024)
+> [!IMPORTANT]
+> This version includes significant updates and improvements.
+
+### Breaking changes
+- `response.cookies` is now a property instead of a method. [Note: `response.cookies()` has been removed and replaced with `response.cookies`.]
+
+### Additional updates
+- Added `session.cookies` property to retrieve all session cookies. For detailed information, see [examples/session-cookies.md](https://github.com/Sahil1337/node-tls-client/tree/main/examples/samples/cookie-example.md).
+- Added `byteResponse` option in requestOptions. Setting this to true ensures that the response is treated as a byte response. For detailed information, see [examples/images.js](https://github.com/Sahil1337/node-tls-client/tree/main/examples/images.js)
+- Added functionality to send binary data. Note: Ensure that the appropriate headers are set before sending binary data in the body.
+- Added `disableIPV6` and `disableIPV4` options in `SessionOptions`. For detailed information, see [SessionOptions](https://sahil1337.github.io/node-tls-client/interfaces/SessionOptions.html#disableIPV4)
+- Added `hostOverride` property in requestOptions. Used to override the Host header, typically needed when making requests directly to an IP address.
+- `chrome_124` profile will be used when no `client_identifier` is specified in session options.
+- Refactored cookies.ts and session typinhd for improved readability (internal change, no impact on existing use cases)
+- Updated TlsResponse according to new changes.
+
 # 1.1.3 (June 13, 2024)
 > [!WARNING]  
 > This version contains breaking changes.